requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==22.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.2
dateparser==1.2.0
decorator==5.1.1
dill==0.3.8
distlib==0.3.6
Django==5.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.2.0
django-filter==23.2
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
drf-spectacular-sidecar==2023.3.1
drf-yasg==1.21.5
executing==2.0.1
Faker==17.6.0
filelock==3.12.2
fonttools==4.51.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
imageio==2.34.0
incremental==22.10.0
inflection==0.5.1
ipython==8.23.0
ipywidgets==8.1.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
jsonschema==4.17.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
MetaTrader5==5.0.4200
#mt5linux
mypy-extensions==1.0.0
numba==0.59.1
numpy==1.26.4
packaging==23.0
pandas==2.2.2
pandas_ta==0.3.14b0
parso==0.8.4
passlib==1.7.4
Pillow==9.4.0
platformdirs==3.8.0
plotly==5.21.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.6.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
regex==2024.4.16
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schedule==1.2.1
scikit-learn==1.4.2
scipy==1.13.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.3
swagger-spec-validator==3.0.3
tenacity==8.2.3
threadpoolctl==3.4.0
toposort==1.10
tqdm==4.66.2
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.15
vectorbt==0.26.1
virtualenv==20.23.1
wcwidth==0.2.13
whitenoise==6.4.0
widgetsnbextension==4.0.10
zope.interface==6.3
